requirements.txt
streamlit
openai
groq
pandas
matplotlib
seaborn
plotly
fpdf2
openpyxl
python-dotenv